implement sqlx for some parts of server dao

+++ src-tauri/src/model/dto/server.rs
@@ -6,7 +6,7 @@ use super::MsqtDto;
 #[derive(Serialize, Deserialize, Clone, Debug, sqlx::FromRow)]
 #[serde(rename_all = "camelCase")]
 pub struct Server {
-    pub(in crate::model) id: u64,
+    pub(in crate::model) id: u32,
     pub(in crate::model) name: String,
     pub(in crate::model) url: String,
     pub(in crate::model) port: u16,
@@ -14,7 +14,7 @@ pub struct Server {
 }
 
 impl MsqtDto for Server {
-    fn id(&self) -> u64 {
+    fn id(&self) -> u32 {
         self.id
     }
 }
