+++ src-tauri/src/model/dto/topic.rs
@@ -6,21 +6,21 @@ use super::MsqtDto;
 #[derive(Serialize, Deserialize, Clone, Debug, sqlx::FromRow)]
 #[serde(rename_all = "camelCase")]
 pub struct Topic {
-    pub(in crate::model) id: u64,
-    pub(in crate::model) fk_server_id: u64,
+    pub(in crate::model) id: u32,
+    pub(in crate::model) fk_server_id: u32,
     pub(in crate::model) name: String,
     pub(in crate::model) enabled: bool,
 }
 
 impl MsqtDto for Topic {
-    fn id(&self) -> u64 {
+    fn id(&self) -> u32 {
         self.id
     }
 }
 
 #[allow(dead_code)]
 impl Topic {
-    pub fn server_id(&self) -> u64 {
+    pub fn server_id(&self) -> u32 {
         self.fk_server_id
     }
     pub fn get_subscribe_filter(&self) -> SubscribeFilter {
